pyproject.toml
[tool.poetry]
name = "airflow-pybr-2021"
version = "0.1.0"
description = "Airflow Python Brasil 2021"
authors = ["Tarsis Azevedo"]


[tool.poetry.dependencies]
python = "^3.8"
apache-airflow = {version = "^2.1.4"}

[tool.poetry.dev-dependencies]
black = "^21.5b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"